IPPO tidy up
