+++ venv_wsl/pyvenv.cfg
@@ -1,0 +1,5 @@
+home = /usr/bin
+include-system-site-packages = false
+version = 3.11.10
+executable = /usr/bin/python3.11
+command = /usr/bin/python3 -m venv /mnt/c/Users/mavritsa/Repositories/jax-agents/venv_wsl
